requirements.txt
Django==4.1.3
djangorestframework==3.14.0
django-filter==22.1
django-admin-sortable2==2.1.4
Pillow==9.3.0
django-debug-toolbar==3.7.0
django-versatileimagefield==2.2
psycopg2-binary==2.9.5
django-ckeditor==6.5.1
drf-yasg==1.21.4
